+++ app/src/main/java/eu/faircode/email/Core.java
@@ -4055,6 +4055,30 @@ class Core {
 
                         message = dup;
                         process = true;
+                    } else if (msgid != null && EntityFolder.DRAFTS.equals(folder.type)) {
+                        try {
+                            if (dup.uid < uid) {
+                                MimeMessage existing = (MimeMessage) ifolder.getMessageByUID(dup.uid);
+                                if (existing != null &&
+                                        msgid.equals(existing.getHeader(MessageHelper.HEADER_CORRELATION_ID, null))) {
+                                    Log.e(folder.name + " late draft" +
+                                            " host=" + account.host + " uid=" + dup.uid + "<" + uid);
+                                    existing.setFlag(Flags.Flag.DELETED, true);
+                                    expunge(context, ifolder, Arrays.asList(existing));
+                                    db.message().setMessageUiHide(dup.id, true);
+                                }
+                            } else if (dup.uid > uid) {
+                                if (msgid.equals(imessage.getHeader(MessageHelper.HEADER_CORRELATION_ID, null))) {
+                                    Log.e(folder.name + " late draft" +
+                                            " host=" + account.host + " uid=" + dup.uid + ">" + uid);
+                                    imessage.setFlag(Flags.Flag.DELETED, true);
+                                    expunge(context, ifolder, Arrays.asList(imessage));
+                                    return null;
+                                }
+                            }
+                        } catch (Throwable ex) {
+                            Log.e(ex);
+                        }
                     }
                 }