+++ app/src/main/java/eu/faircode/email/FragmentOptionsMisc.java
@@ -165,10 +165,12 @@ public class FragmentOptionsMisc extends FragmentBase implements SharedPreferenc
     private TextView tvRoomQueryThreads;
     private SeekBar sbRoomQueryThreads;
     private ImageButton ibRoom;
+    private SwitchCompat swIntegrity;
     private SwitchCompat swWal;
     private SwitchCompat swCheckpoints;
     private SwitchCompat swAnalyze;
     private SwitchCompat swAutoVacuum;
+    private SwitchCompat swSyncExtra;
     private TextView tvSqliteCache;
     private SeekBar sbSqliteCache;
     private TextView tvChunkSize;
@@ -238,7 +240,8 @@ public class FragmentOptionsMisc extends FragmentBase implements SharedPreferenc
             "watchdog", "experiments", "main_log", "protocol", "log_level", "debug", "leak_canary",
             "test1", "test2", "test3", "test4", "test5",
             "work_manager", // "external_storage",
-            "query_threads", "wal", "sqlite_checkpoints", "sqlite_analyze", "sqlite_auto_vacuum", "sqlite_cache",
+            "query_threads",
+            "sqlite_integrity_check", "wal", "sqlite_checkpoints", "sqlite_analyze", "sqlite_auto_vacuum", "sqlite_sync_extra", "sqlite_cache",
             "chunk_size", "thread_range", "undo_manager",
             "webview_legacy", "browser_zoom", "fake_dark",
             "show_recent",
@@ -268,7 +271,8 @@ public class FragmentOptionsMisc extends FragmentBase implements SharedPreferenc
             "eml_auto_confirm",
             "open_with_pkg", "open_with_tabs",
             "gmail_checked", "outlook_checked",
-            "redmi_note"
+            "redmi_note",
+            "accept_space", "accept_unsupported"
     };
 
     @Override
@@ -362,10 +366,12 @@ public class FragmentOptionsMisc extends FragmentBase implements SharedPreferenc
         tvRoomQueryThreads = view.findViewById(R.id.tvRoomQueryThreads);
         sbRoomQueryThreads = view.findViewById(R.id.sbRoomQueryThreads);
         ibRoom = view.findViewById(R.id.ibRoom);
+        swIntegrity = view.findViewById(R.id.swIntegrity);
         swWal = view.findViewById(R.id.swWal);
         swCheckpoints = view.findViewById(R.id.swCheckpoints);
         swAnalyze = view.findViewById(R.id.swAnalyze);
         swAutoVacuum = view.findViewById(R.id.swAutoVacuum);
+        swSyncExtra = view.findViewById(R.id.swSyncExtra);
         tvSqliteCache = view.findViewById(R.id.tvSqliteCache);
         sbSqliteCache = view.findViewById(R.id.sbSqliteCache);
         ibSqliteCache = view.findViewById(R.id.ibSqliteCache);
@@ -1102,6 +1108,17 @@ public class FragmentOptionsMisc extends FragmentBase implements SharedPreferenc
             }
         });
 
+        swIntegrity.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
+            @Override
+            public void onCheckedChanged(CompoundButton v, boolean checked) {
+                prefs.edit()
+                        .putBoolean("sqlite_integrity_check", checked)
+                        .remove("debug")
+                        .commit();
+                ApplicationEx.restart(v.getContext(), "sqlite_integrity_check");
+            }
+        });
+
         swWal.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
             @Override
             public void onCheckedChanged(CompoundButton compoundButton, boolean checked) {
@@ -1129,8 +1146,19 @@ public class FragmentOptionsMisc extends FragmentBase implements SharedPreferenc
                 prefs.edit()
                         .putBoolean("sqlite_auto_vacuum", checked)
                         .remove("debug")
-                        .apply();
+                        .commit();
                 ApplicationEx.restart(v.getContext(), "sqlite_auto_vacuum");
+            }
+        });
+
+        swSyncExtra.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
+            @Override
+            public void onCheckedChanged(CompoundButton v, boolean checked) {
+                prefs.edit()
+                        .putBoolean("sqlite_sync_extra", checked)
+                        .remove("debug")
+                        .commit();
+                ApplicationEx.restart(v.getContext(), "sqlite_sync_extra");
             }
         });
 
@@ -1970,10 +1998,12 @@ public class FragmentOptionsMisc extends FragmentBase implements SharedPreferenc
         tvRoomQueryThreads.setText(getString(R.string.title_advanced_room_query_threads, NF.format(query_threads)));
         sbRoomQueryThreads.setProgress(query_threads);
 
+        swIntegrity.setChecked(prefs.getBoolean("sqlite_integrity_check", true));
         swWal.setChecked(prefs.getBoolean("wal", true));
         swCheckpoints.setChecked(prefs.getBoolean("sqlite_checkpoints", true));
         swAnalyze.setChecked(prefs.getBoolean("sqlite_analyze", true));
-        swAutoVacuum.setChecked(prefs.getBoolean("sqlite_auto_vacuum", !Helper.isRedmiNote()));
+        swAutoVacuum.setChecked(prefs.getBoolean("sqlite_auto_vacuum", false));
+        swSyncExtra.setChecked(prefs.getBoolean("sqlite_sync_extra", true));
 
         int sqlite_cache = prefs.getInt("sqlite_cache", DB.DEFAULT_CACHE_SIZE);
         Integer cache_size = DB.getCacheSizeKb(getContext());