+++ app/src/main/java/eu/faircode/email/Helper.java
@@ -174,7 +174,7 @@ public class Helper {
     static final float LOW_LIGHT = 0.6f;
 
     static final int BUFFER_SIZE = 8192; // Same as in Files class
-    static final long MIN_REQUIRED_SPACE = 250 * 1024L * 1024L;
+    static final long MIN_REQUIRED_SPACE = 100 * 1000L * 1000L;
     static final int AUTOLOCK_GRACE = 15; // seconds
     static final long PIN_FAILURE_DELAY = 3; // seconds
 
@@ -888,7 +888,7 @@ public class Helper {
 
     static void view(Context context, Intent intent) {
         Uri uri = intent.getData();
-        if ("http".equals(uri.getScheme()) || "https".equals(uri.getScheme()))
+        if (UriHelper.isHyperLink(uri))
             view(context, intent.getData(), false);
         else
             try {
@@ -924,7 +924,9 @@ public class Helper {
                 " isInstalled=" + isInstalled(context, open_with_pkg) +
                 " hasCustomTabs=" + hasCustomTabs(context, uri, open_with_pkg));
 
-        if (!UriHelper.isHyperLink(uri)) {
+        if (UriHelper.isHyperLink(uri))
+            uri = UriHelper.fix(uri);
+        else {
             open_with_pkg = null;
             open_with_tabs = false;
         }
@@ -1100,7 +1102,7 @@ public class Helper {
     }
 
     static Intent getIntentIssue(Context context, String reference) {
-        if (ActivityBilling.isPro(context)) {
+        if (ActivityBilling.isPro(context) && false) {
             String version = BuildConfig.VERSION_NAME + BuildConfig.REVISION + "/" +
                     (Helper.hasValidFingerprint(context) ? "1" : "3") +
                     (BuildConfig.PLAY_STORE_RELEASE ? "p" : "") +
@@ -2147,6 +2149,21 @@ public class Helper {
 
     static boolean isDot(char c) {
         return (c == '.' /* Latin */ || c == '。' /* Chinese */);
+    }
+
+    static String trim(String value, String chars) {
+        if (value == null)
+            return null;
+
+        for (Character kar : chars.toCharArray()) {
+            String k = kar.toString();
+            while (value.startsWith(k))
+                value = value.substring(1);
+            while (value.endsWith(k))
+                value = value.substring(0, value.length() - 1);
+        }
+
+        return value;
     }
 
     // Files