+++ push.sh
@@ -3,5 +3,5 @@
 for remote in $(git remote | grep -v origin)
 do
     echo "Pushing to $remote";
-    git push "$remote" --all;
+    git push "$remote" --all $@;
 done